+++ index.html
@@ -106,6 +106,8 @@
     <script src="src/storage.js"></script>
     <script src="src/permission.js"></script>
     <script src="src/map.js"></script>
+    <script src="src/data-layer.js"></script>
+    <script src="src/ui-controller.js"></script>
     <script src="main.js"></script>
-</body>
+  </body>
 </html>
